Minor tweaks as I find them...
Aside: I would also think about the rm -r in `rm_if_exists`, but that's only my paranoia as I know `mktemp` is using a path, but I like to cd to a well know directory path and then only rm -r the named remainder of a custom path. (Perhaps I would be worried someone would change the tmp path (i.e. tmpDir) and/or get it wrong and wipe out some parent structure, but I have not edited/suggested the fix in the script for that yet.